moved some  code from AIUtil to ClientAI
